Updates both POSTs to explicitly declare Form Post
The AJAX `datatype` key threw me off when looking at this example as I had assumed that was the content type of the POST request, which it is not. AJAX POSTs are implicitly `application/x-www-form-urlencoded` which is an important thing to know as part of this flow. Making this explicit will hopefully avoid others from running into the same problem that I did. 

Full details @ [#639](https://github.com/FusionAuth/fusionauth-issues/issues/639).